# docs/openapi/moderation-suppliers.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Barello Moderation Suppliers API",
    "version": "1.0.0"
  },
  "paths": {
    "/moderation/suppliers": {
      "get": {
        "operationId": "getModerationSuppliers",
        "responses": {
          "200": {
            "description": "Supplier list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModerationSupplier"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createModerationSupplier",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "inn",
                  "companyName",
                  "ownerPhone",
                  "companyPhone",
                  "city",
                  "address",
                  "accessStatus"
                ],
                "properties": {
                  "inn": {
                    "type": "string",
                    "pattern": "^\\d{10}(\\d{2})?$"
                  },
                  "companyName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "ownerFullName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "ownerPhone": {
                    "type": "string",
                    "pattern": "^\\+7\\d{10}$"
                  },
                  "companyPhone": {
                    "type": "string",
                    "pattern": "^\\+7\\d{10}$"
                  },
                  "city": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100
                  },
                  "address": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 300
                  },
                  "accessStatus": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "BLOCKED"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Supplier created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "ok",
                    "supplier"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "supplier": {
                      "$ref": "#/components/schemas/ModerationSupplier"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Request failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ModerationSupplier": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "publicId",
          "tableId",
          "businessId",
          "inn",
          "name",
          "displayName",
          "companyName",
          "catalogName",
          "city",
          "address",
          "contactName",
          "ownerFullName",
          "ownerName",
          "ownerPhone",
          "owner",
          "hasActiveAccount",
          "accountStatus",
          "membershipId",
          "membershipStatus",
          "membershipAccessLevel",
          "membershipJoinedAt",
          "loginPhone",
          "companyPhone",
          "phone",
          "email",
          "website",
          "accessStatus",
          "lastLoginAt",
          "categoriesCount",
          "supplierProductsCount",
          "productsCount",
          "offersCount",
          "promoCount",
          "priceImportsCount",
          "lastPriceImportAt",
          "lastPriceImport",
          "isActive",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "publicId": {
            "type": "string"
          },
          "tableId": {
            "type": "string"
          },
          "businessId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "inn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "catalogName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "contactName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "ownerFullName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "ownerName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "ownerPhone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "owner": {
            "anyOf": [
              {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "id",
                  "publicId",
                  "fullName",
                  "loginPhone",
                  "status",
                  "accountType",
                  "lastLoginAt"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "publicId": {
                    "type": "string"
                  },
                  "fullName": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "loginPhone": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "UNIDENTIFIED",
                      "ACTIVE",
                      "BLOCKED",
                      "ARCHIVED"
                    ]
                  },
                  "accountType": {
                    "type": "string",
                    "enum": [
                      "VENUE_STAFF",
                      "SUPPLIER_STAFF",
                      "UNIDENTIFIED"
                    ]
                  },
                  "lastLoginAt": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "date-time"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                }
              },
              {
                "type": "null"
              }
            ]
          },
          "hasActiveAccount": {
            "type": "boolean"
          },
          "accountStatus": {
            "type": "string"
          },
          "membershipId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "membershipStatus": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PENDING",
                  "ACTIVE",
                  "REJECTED",
                  "REVOKED"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "membershipAccessLevel": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ADMIN",
                  "SENIOR_STAFF",
                  "LINE_STAFF"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "membershipJoinedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "loginPhone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "companyPhone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "accessStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BLOCKED"
            ]
          },
          "lastLoginAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "categoriesCount": {
            "type": "integer",
            "minimum": 0
          },
          "supplierProductsCount": {
            "type": "integer",
            "minimum": 0
          },
          "productsCount": {
            "type": "integer",
            "minimum": 0
          },
          "offersCount": {
            "type": "integer",
            "minimum": 0
          },
          "promoCount": {
            "type": "integer",
            "minimum": 0
          },
          "priceImportsCount": {
            "type": "integer",
            "minimum": 0
          },
          "lastPriceImportAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "lastPriceImport": {
            "anyOf": [
              {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "id",
                  "fileName",
                  "status",
                  "createdAt",
                  "rowsCount",
                  "failedRows",
                  "errorText"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "fileName": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "UPLOADED",
                      "QUEUED",
                      "PROCESSING",
                      "PARSED",
                      "HAS_ISSUES",
                      "READY_TO_PUBLISH",
                      "PROCESSED",
                      "PUBLISHED",
                      "PARTIALLY_PUBLISHED",
                      "FAILED",
                      "CANCELLED"
                    ]
                  },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "rowsCount": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "failedRows": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "errorText": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                }
              },
              {
                "type": "null"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}